refactor(cdn): extract cdn utils into separate module
